Preliminary update to protocols and delegates
