+++ Jasmine/Common/ViewControllers/Delegates/GameStatusUpdateDelegate.swift
@@ -1,0 +1,7 @@
+protocol GameStatusUpdateDelegate: class {
+
+    // MARK: Game Status
+    /// Tells the implementor of the delegate that the game status has been updated.
+    /// - Note: This method can be used to update the game status that is displayed on views.
+    func gameStatusDidUpdate()
+}
