+++ Jasmine/Common/ViewControllers/Delegates/TimeUpdateDelegate.swift
@@ -1,0 +1,6 @@
+protocol TimeUpdateDelegate: class {
+
+    /// Tells the implementor of the delegate that the time has been updated.
+    /// - Note: This method can be used to update the remaining time that is displayed on views.
+    func timeDidUpdate()
+}
