--- Jasmine/Common/ViewControllers/GameElements/ScoreUpdateDelegate.swift
@@ -1,3 +1,0 @@
-protocol ScoreUpdateDelegate: class {
-    func scoreDidUpdate()
-}
