--- Jasmine/Common/ViewControllers/GameElements/TimeUpdateDelegate.swift
@@ -1,3 +1,0 @@
-protocol TimeUpdateDelegate: class {
-    func timeDidUpdate()
-}
