+++ Jasmine/Common/ViewModels/Protocols/TimeableProtocol.swift
@@ -1,9 +1,14 @@
 import Foundation
 
-/// A ViewModel that implements a timer, i.e. it is timed.
-protocol CountdownTimable {
+protocol TimeableProtocol {
+
+    /// A delegate that notifies the implementing view controller that the time (remaining) has been
+    /// updated.
+    weak var timeDelegate: TimeUpdateDelegate? { get }
+
     /// Specifies the total time allowed in the game.
     var totalTimeAllowed: TimeInterval { get }
+
     /// Specifies the remaining time left in the game.
     var timeRemaining: TimeInterval { get }
 }
