--- Jasmine/Flappy/ViewControllers/FlappyViewControllerDelegate.swift
@@ -1,7 +1,0 @@
-import Foundation
-
-/// Implement this delegate for `FlappyViewModelProtocol` to call for updating and commanding the
-/// implementing game view controller.
-protocol FlappyViewControllerDelegate: BaseGameViewControllerDelegate {
-
-}
